# packages/backend/openapi.yml
openapi: 3.0.3
info:
  title: Bid Eligibility Judgment System API
  version: 0.1.0
  description: Backend API for bid evaluation, announcements, partners, orderers, contacts, and companies.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://bidapp-backend-postgres-50843898931.asia-northeast1.run.app
    description: Cloud Run production
paths:
...
  /api/evaluations:
    get:
      summary: List evaluations with filters and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 25
            maximum: 1000
        - name: statuses
          in: query
          schema:
            type: array
            items:
              type: string
        - name: workStatuses
          in: query
          schema:
            type: array
            items:
              type: string
        - name: searchQuery
          in: query
          schema:
            type: string
        - name: sortField
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated evaluation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Evaluation'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
  /api/evaluations/{id}:
    get:
      summary: Get evaluation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Single evaluation
        '404':
          description: Not found
...
  /api/announcements/{announcementNo}:
    get:
      summary: Get announcement by number
      parameters:
        - name: announcementNo
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Single announcement
        '404':
          description: Not found
...
components:
  schemas:
    Evaluation:
      type: object
      properties:
        id:
          type: string
        evaluationNo:
          type: string
        announcement:
          type: object
        company:
          type: object
        branch:
          type: object
        status:
          type: string
          enum: [all_met, other_only_unmet, unmet]
        workStatus:
          type: string
          enum: [not_started, in_progress, completed]
        currentStep:
          type: integer
        evaluatedAt:
          type: string
...